ES-016.3b PayU refactor code
