+++ src/main/java/com/wojciechbarwinski/demo/epic_board_games_shop/payments/payU/PayUPaymentAdapter.java
@@ -5,6 +5,9 @@ import com.wojciechbarwinski.demo.epic_board_games_shop.entities.Order;
 import com.wojciechbarwinski.demo.epic_board_games_shop.payments.PaymentDataDTO;
 import com.wojciechbarwinski.demo.epic_board_games_shop.payments.PaymentPort;
 import com.wojciechbarwinski.demo.epic_board_games_shop.payments.payU.dtos.PayUOrderRequestDTO;
+import com.wojciechbarwinski.demo.epic_board_games_shop.payments.payU.utils.PayUAuthTokenProvider;
+import com.wojciechbarwinski.demo.epic_board_games_shop.payments.payU.utils.PayUNewOrderProvider;
+import com.wojciechbarwinski.demo.epic_board_games_shop.payments.payU.utils.PayUOrderCreator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
